Refactor common query logic into extensions
